Use a macro to implement context methods
This attempts to solve the issue of duplicate code and out of
date documentation for methods shared between multiple contexts.

The approach taken here is to use a simple macro to implement
those methods.

The macro was actually a real pain to write, and kind of gross API
wise; in particular you need to pass in all of the lifetimes twice.
I've given up trying to improve it, and given that this is private API
it doesn't matter too much, but it does offend my sensibilites
somewhat.

There are a few bits of future work; window & window_id should
be shared between all contexts, but that requires a change to
PaintCtx; another change would let us reuse is_focused.

In any case, I think this is a step in a good direction.